+++ R/auth.R
@@ -162,13 +162,13 @@ rtweet_app <- function(bearer_token) {
   )
 }
 
-ask_pass <- function(type, call = caller_env()) {
-  check_installed("askpass", call = call)
+ask_pass <- function(type) {
+  check_installed("askpass", call = current_call())
 
   message <- paste0("Please enter your ", type, ": ")
   val <- askpass::askpass(message)
   if (is.null(val)) {
-    abort("Cancelled by user")
+    abort("Cancelled by user", call = current_call())
   }
   val
 }
@@ -286,7 +286,7 @@ auth_as <- function(auth = NULL) {
   invisible(old)
 }
 
-find_auth <- function(auth = NULL, call = caller_env()) {
+find_auth <- function(auth = NULL) {
   if (is.null(auth)) {
     if (is_developing()) {
       rtweet_test() %||% no_token()
@@ -296,29 +296,33 @@ find_auth <- function(auth = NULL, call = caller_env()) {
   } else if (is_auth(auth)) {
     auth
   } else if (is_string(auth)) {
-    load_token(auth, call)
+    load_token(auth)
   } else {
     abort("Unrecognised input to `auth`",
-          call = call)
+          call = current_call())
   }
 }
 
-load_token <- function(auth_name, call = caller_env) {
+load_token <- function(auth_name) {
   if (file.exists(auth_name)) {
     path <- auth_name
+    name <- tools::file_path_sans_ext(basename(auth_name))
   } else {
     path <- auth_path(paste0(auth_name, ".rds"))
+    name <- auth_name
   }
   if (!file.exists(path)) {
     abort(paste0("Can't find saved auth with name '", auth_name, "'"),
-          call = call)
+          call = current_call())
   }
   if (!is_developing()) {
     inform(paste0("Reading auth from '", path, "'"))
   }
-  readRDS(path)
+  token <- readRDS(path)
+  # Store the name of the token as an attribute
+  attr(token, "name") <- name
+  token
 }
-
 
 default_cached_auth <- function() {
   default <- auth_path("default.rds")
@@ -329,12 +333,13 @@ default_cached_auth <- function() {
 
   names <- auth_list()
   if (length(names) == 0) {
-    abort("No default authentication found. Please call `auth_setup_default()`")
+    abort("No default authentication found. Please call `auth_setup_default()`",
+          call = caller_call())
   } else {
     abort(c(
       "No default authentication found. Pick existing auth with:",
       paste0("auth_as('", names, "')")
-    ))
+    ), call = caller_call())
   }
 }
 
@@ -344,11 +349,11 @@ auth_has_default <- function() {
   file.exists(auth_path("default.rds"))
 }
 
-no_token <- function(call = caller_env()) {
+no_token <- function() {
   if (is_testing()) {
     testthat::skip("Auth not available")
   } else {
-    abort("Could not authenticate", call = call)
+    abort("Could not authenticate", call = current_call())
   }
 }
 # Internal function to generate the bot used for testing
@@ -421,12 +426,14 @@ rtweet_oauth2 <- function(client = NULL, scopes = NULL) {
   client <- client %||% client_get()
   if (!is_client(client)) {
     abort(c("Client not valid",
-          i = "Check out the `vignette('auth', 'rtweet')`."))
+          i = "Check out the `vignette('auth', 'rtweet')`."),
+          call = current_call())
   }
   if (is.null(scopes)) {
     scopes <- client_scopes(client)
   } else if (!check_scopes(scopes)) {
-    abort("Scopes is not in the right format.")
+    abort("Scopes is not in the right format.",
+          call = current_call())
   }
 
   # Guide to all urls for OAuth 2
@@ -444,7 +451,7 @@ rtweet_oauth2 <- function(client = NULL, scopes = NULL) {
                        )
   # Example of valid url for authorization (created via client_as("academic_dev");rtweet_oauth2() )
   # https://twitter.com/i/oauth2/authorize?response_type=code&client_id=Tm5FWFA3OGFUVUMxTjhUREwzZzQ6MTpjaQ&redirect_uri=http%3A%2F%2F127.0.0.1%3A1410%2F&scope=tweet.read%20tweet.write%20tweet.moderate.write%20users.read%20follows.read%20follows.write%20offline.access%20space.read%20mute.read%20mute.write%20like.read%20like.write%20list.read%20list.write%20block.read%20block.write%20bookmark.read%20bookmark.write&state=PVgWK3MviQ5MBsfj0Iy5D89HBFR4mPwTl0yumjSPlWo&code_challenge=FNcGvupIzNIbWOL8rdJOrxsVS_b2R01vIbynF_iQIMQ&code_challenge_method=S256
-  # # Note that the cliend_id should match in the url
+  # # Note that the client_id should match in the url
   inform("Requires confirming permissions to the app (client) every two hours!")
   attr(token, "app") <- attr(client, "app")
   token
@@ -452,17 +459,11 @@ rtweet_oauth2 <- function(client = NULL, scopes = NULL) {
 
 # Renew token if needed
 # Makes the assumption that the right app is still in the user computer
-auth_renew <- function(token, scopes = NULL, call = caller_env()) {
+auth_renew <- function(token, scopes = NULL) {
   stopifnot(auth_is_pkce(token))
 
   if (.POSIXct(token$expires_at) >= Sys.time()) {
     return(token)
-  }
-  msg <- "Impossible to renew the authentication without interactive usage."
-  if (!interactive()  && is_testing()) {
-    testthat::skip(msg)
-  } else if (!interactive()  && !is_developing()) {
-    abort(msg)
   }
 
   if (!is.null(scopes) && check_scopes(scopes)) {
@@ -470,38 +471,43 @@ auth_renew <- function(token, scopes = NULL, call = caller_env()) {
   } else if (is.null(scopes)) {
     scopes <- strsplit(token$scope, " ", fixed = TRUE)[[1]]
   } else {
-    abort("Scopes is not in the right format.")
+    abort("Scopes is not in the right format.",
+          call = current_call())
   }
   client_name <- attr(token, "app", TRUE)
   client <- load_client(client_name)
 
-  # TODO: once it is possible to silently update the token (see https://github.com/r-lib/httr2/issues/197)
-  # activate this branch to reauthorize the app
-  if (FALSE ) {
-    !any(grepl("offline.access", scopes, fixed = TRUE))
-    abort(c("Not possible to update the client"))
-    token <- rtweet_oauth2(client, scopes)
-    return(token)
-  }
-
-  # inform("You'll need to give again permissions to the app every two hours!")
-  token <- rtweet_oauth2(client, scopes)
-  # The provided refresh token can only be used once:
-  # https://twittercommunity.com/t/unable-to-obtain-new-access-token-by-using-refresh-token/164123/16
-  # token <- httr2::oauth_flow_refresh(client, refresh_token = token$refresh_token,
-  #                                scope = token$scopes)
-  # This is reported to upstream: https://github.com/r-lib/httr2/issues/197
-
+  # Is possible to silently update the token
+  # see https://github.com/r-lib/httr2/issues/197
+  token2 <- httr2::oauth_flow_refresh(client,
+                                      refresh_token = token$refresh_token,
+                                      scope = token$scope)
   # Save token in the environment
-  # It could be that token is not from the environment, but this feel safer than saving it in a file directly.
-  auth_as(token)
-  inform(c("i" = "Using renewed token",
-           "!" = "Save your new oauth2 token with the appropriate name:",
-           "`auth_save(auth_get(), 'renewed_token')`"))
-  token
+  auth_as(token2)
+  # If possible replace it in the user storage.
+  resave_token(token2, token_name(token))
+  token2
 }
 
 decrypt <- function(x) {
   check_installed("openssl") # Suggested!
   rawToChar(openssl::rsa_decrypt(x[[2]], x[[1]]))
+}
+
+token_name <- function(x){
+  attr(x, "name", TRUE)
+}
+
+resave_token <- function(token, name) {
+  name <- token_name(token)
+  if (is.null(name)) {
+    inform(c("!" = "It was not possible to save your token",
+             "i" = "Saving your new oauth2 token with: `auth_save(auth_get(), 'renewed_token')`"))
+    auth_save(auth_get(), 'renewed_token')
+  }
+
+  path <- auth_path(paste0(name, ".rds"))
+  # Protects in case the folder gets deleted while the token is loaded
+  dir.create(auth_path(), showWarnings = FALSE, recursive = TRUE)
+  saveRDS(token, path)
 }