+++ tests/testthat/test-req_errors.R
@@ -1,5 +1,5 @@
 test_that("req_error works", {
-  testing_with_authentication("renewed_token")
+  testing_with_authentication("bearer_testing_app")
   expect_error(user_blocked("407200271", n = Inf),
                "must be the same as the authenticating user")
 })