+++ NEWS.md
@@ -1,4 +1,4 @@
-# rtweet (development version)
+# rtweet 1.2.1
 
 * Fix `auth_sitrep()` to work well with OAuth2 tokens.